+++ CHANGELOG.rst
@@ -1,4 +1,4 @@
-HEAD (2024-??-??)
+0.35.0 (2024-09-10)
 -------------------
 
 *	Add new UL4 attributes ``search_url``, ``search_param_name`` and ``target_param_name`` in
@@ -25,7 +25,7 @@ HEAD (2024-??-??)
 *	Now each vSQL rule stores the specification it was generated from. A string
 	version of that can be retrieved via the method :meth:`str_vsqlsource`.
 
-*	Added the methods :meth:`App.send_mail` and :meth:`App.send_mail`.
+*	Added the methods :meth:`App.send_mail` and :meth:`Record.send_mail`.
 
 
 0.32.0 (2024-06-14)